testdata/htmltodo/box-sizing-invalid.htm

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>CSS UI Level 3: parsing box-sizing with invalid values</title>
<link href="mailto:ericwilligers@chromium.org" rel="author" title="Eric Willigers">
<link href="https://drafts.csswg.org/css-ui-3/#box-sizing" rel="help">
<meta content="box-sizing supports only the grammar 'content-box | border-box'." name="assert">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("box-sizing", "auto");
test_invalid_value("box-sizing", "content-box border-box");
test_invalid_value("box-sizing", "fill-box");
test_invalid_value("box-sizing", "margin-box");
test_invalid_value("box-sizing", "padding-box");
test_invalid_value("box-sizing", "view-box");
</script>


</body></html>
